Split worker to several files
